reorganize file structure, use service instead of class for image formating
