+++ src/app/app.module.ts
@@ -4,10 +4,11 @@ import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
 
 import { AppComponent } from './app.component';
-import { EditorComponent } from './editor/editor.component';
-import { UploaderComponent } from './uploader/uploader.component';
-import { PickerButtonsComponent } from './uploader/picker-buttons/picker-buttons.component';
-import { FirstUppercasePipe } from './uploader/picker-buttons/first-uppercase.pipe';
+import { EditorComponent } from './views/editor/editor.component';
+import { UploaderComponent } from './views/uploader/uploader.component';
+import { PickerButtonsComponent } from './views/uploader/picker-buttons/picker-buttons.component';
+import { FirstUppercasePipe } from './views/uploader/picker-buttons/first-uppercase.pipe';
+import { ImageService } from './services/image-service/image.service';
 
 const routes: Routes = [
   { path: '', redirectTo: '/uploader', pathMatch: 'full' },
@@ -31,7 +32,7 @@ const routes: Routes = [
       { enableTracing: true } // <-- debugging purposes only
     )
   ],
-  providers: [],
-  bootstrap: [AppComponent]
+  bootstrap: [AppComponent],
+  providers: [ImageService]
 })
 export class AppModule { }
