+++ src/app/services/image-service/image.service.spec.ts
@@ -1,0 +1,15 @@
+import { TestBed, inject } from '@angular/core/testing';
+
+import { ImageService } from './image.service';
+
+describe('ImageService', () => {
+  beforeEach(() => {
+    TestBed.configureTestingModule({
+      providers: [ImageService]
+    });
+  });
+
+  it('should be created', inject([ImageService], (service: ImageService) => {
+    expect(service).toBeTruthy();
+  }));
+});
