+++ src/app/views/editor/editor.component.css
