+++ src/app/views/editor/editor.component.spec.ts
