+++ src/app/views/uploader/picker-buttons/first-uppercase.pipe.spec.ts
