+++ src/app/views/uploader/picker-buttons/picker-buttons.component.css
