+++ src/app/views/uploader/picker-buttons/picker-buttons.component.html
@@ -2,8 +2,8 @@
   <button *ngFor="let button of buttons" class="btn btn-light" [id]="button.name" (click)="uploadImage(button.name)"
     [ngClass]="{'disabled': button.uploaded, 'left': button.checkPosition === 'left', 'right': button.checkPosition === 'right'}">
     Pick {{ button.name | firstUppercase }}
-    <img *ngIf="button.uploaded" (click)="delete(button.name)" class="cross_icon" src="../../../assets/cross.png">
-    <img *ngIf="button.uploaded" class="check_icon" src="../../../assets/check.svg"
+    <img *ngIf="button.uploaded" (click)="delete(button.name)" class="cross_icon" src="../../../../assets/cross.png">
+    <img *ngIf="button.uploaded" class="check_icon" src="../../../../assets/check.svg"
       [ngClass]="button.checkPosition === 'left' ? 'left_check' : 'right_check'">
   </button>
 </div>
