+++ src/app/views/uploader/uploader.component.css
