+++ src/app/views/uploader/uploader.component.spec.ts
