+++ src/app/views/uploader/uploader.component.ts
